Write cylindrical force-free problem
The precision with the cylindrical problem is not that great, it finds
an equilibrium that is 1-2% different from the input regardless of
resolution. This could be due to boundary effects.

Also add the advance_poloidal_field option. Disabling it allows problems
like the 1D Michel solution to continue working even though they specify
the poloidal field directly instead of using a vector potential. When
set to zero, the code skips the advance_vector_potential step.