Improve robustness of boosted ABC setup
This was crashing before because of doing c2p on guard zones. I added a
cell_type flag to mark zones as interior or guard, but it only works
when ng_col is 0 or 1. Now the whole mesh moves and nothing crashes, but
this setup highlights the need to assign vertex velocities more
sensibly. When the boost is along columns, cells get deformed because
verts on opposite sides of the wall move away from one another.