Put in mesh generation for Narayan's self-similar disk wind
This commit specializes to the paraboloidal case where nu=1. Cells
vector potential data is presently assigned the value of the stream
function, which is only referenced for plotting. Verts have the real
vector potential, which is used by the scheme.

The mesh generation tries to build faces aligned with the local flux
coordinates. The first several cells above the "disk surface" may not
have lateral neighbors. Not sure yet how that will be dealt with, just
keep it in mind.