Get the fluxes to be constant in disk wind problem
Mass, energy, and angular momentum are constant, per unit flux, in the
poloidal direction, for the Blandford-Payne setup. The fluid velocity
had to be set to v = E x B / B^2 to get constant energy flux. Gas
entropy and density obviously need to be very small. Even then we are
still only accurate to ~1% in the fluxes, though that may have more to
do with the mesh geometry than not being exactly force-free.