Work on mesh generation - log spacing on parabolas
Keeping the cell aspect ratios reasonably uniform requires a different
number of cells for each track. So now the generate_vertex stage needs
to query the mesh placement function for the number of verts to expect
for that row, before it iterates over the vertices. This could be mode
more elegant, for example by simply returning a 1 if more verts need to
be paced and a 0 otherwise. Then vertex generation acts like an
iterator, so it should get a state variable.