Begin allowing arbitrary flux function
The goal is for mesh rows to be laid out along field lines, and for the
source terms to be expressed in terms of the derivatives of B and R in
psi and chi. Currently, only the chi direction (distance along the field
line) is resolved. Numerical derivatives on the automatically generated
mesh are validated in the case of monopole and parabloloidal fields.