Fix typos breaking narayan07
These included not using the poloidal field and velocity in the initial
data function, but also using the wrong geometry element for magnetic
flux. It should the area_element, not line_element. It seems that line
element is not in use by the code now. It would be good to re-derive the
Faraday's law step to see why I had it wrong.